refactor: extract shared ad form
